Move conversion from SqlCreateTable to CreateTableOperation to a helper class
